fix(syncuserservice-fix-for-named-arraykey)
C# will interpret the serialized as JSON named array as a type definition. So this means ['admin' : {...} ] will require C# to derserialize with a class of type 'admin' which we do not want. I added an integer based index and provided a solution for the previously implemented method to prevent duplication of users in the array.